+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG for Sulu
 ==================
 
+* dev-master
+    * HOTFIX      #3451 [ContentBundle]           Index correct language of shadow document
+
 * 1.6.1 (2017-07-10)
     * HOTFIX      #3437 [AdminBundle]             Added save url to storage for update-url event
     * HOTFIX      #3433 [ContentBundle]           Fixed url generation for domain root page